+++ unstructured_inference/inference/layout.py
@@ -663,7 +663,7 @@ def parse_ocr_data(ocr_data: dict) -> List[TextRegion]:
         (x1, y1, x2, y2) = l, t, l + w, t + h
         text = ocr_data["text"][i]
         if text:
-            text_region = TextRegion(x1, y1, x2, y2, text)
+            text_region = TextRegion(x1, y1, x2, y2, "OCR", text)
             text_regions.append(text_region)
 
     return text_regions